+++ EventSub/TerminalDisable.cs
@@ -1,0 +1,18 @@
+﻿using OpenLib.Menus;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace suitsTerminal.EventSub
+{
+    internal class TerminalDisable
+    {
+        internal static void OnTerminalDisable()
+        {
+            if (suitsTerminal.OpenBodyCams)
+                OpenBodyCams.ResidualCamsCheck();
+
+
+        }
+    }
+}