+++ Game Patch Stuff/PatchStuff.cs
@@ -6,10 +6,7 @@ using UnityEngine;
 using System.Text;
 using GameNetcodeStuff;
 using static suitsTerminal.Misc;
-using static suitsTerminal.StringStuff;
-using static suitsTerminal.AllSuits;
 using BepInEx.Bootstrap;
-using OpenBodyCams;
 
 namespace suitsTerminal
 {
@@ -34,52 +31,6 @@ namespace suitsTerminal
                     return;
             }
         }
-
-        [HarmonyPatch(typeof(Terminal), "Awake")]
-        public class AwakeTermPatch : Terminal
-        {
-            static void Postfix(Terminal __instance)
-            {
-                suitsTerminal.Terminal = __instance;
-                suitsTerminal.X($"Setting suitsTerminal.Terminal");
-                ResetVars();      
-            }
-
-            private static void ResetVars()
-            {
-                hasLaunched = false;
-                normSuit = 0;
-                showSuit = 0;
-                hintOnce = false;
-                rackSituated = false;
-                favSuitsSet = false;
-                PictureInPicture.PiPCreated = false;
-                suitsTerminal.X("set initial suits values");
-            }
-        }
-
-        [HarmonyPatch(typeof(Terminal), "LoadNewNodeIfAffordable")]
-        public class AffordableNode : Terminal
-        {
-            static void Postfix()
-            {
-                suitsTerminal.X("purchase detected");
-                InitThisPlugin.InitSuitsTerm();
-            }
-        }
-
-        //RemoveLockedSuits
-
-
-        [HarmonyPatch(typeof(StartOfRound), "Start")]
-        public class StartofRoundPatch : MonoBehaviour
-        {
-            static void Postfix()
-            {
-                //old stuff was here
-            }
-        }
-
     }
 
     public class SuitInfo : MonoBehaviour
@@ -119,7 +70,6 @@ namespace suitsTerminal
             {
                 suitsTerminal.X("OpenBodyCams by Zaggy1024 detected!");
                 suitsTerminal.OpenBodyCams = true;
-                GetResolutionForOBC();
             }
             if (Chainloader.PluginInfos.ContainsKey("darmuh.TerminalStuff"))
             {
@@ -127,39 +77,9 @@ namespace suitsTerminal
                 suitsTerminal.TerminalStuff = true;
             }
         }
-
-        private static void GetResolutionForOBC()
-        {
-            List<string> resolutionStrings = GetListFromConfigItem(SConfig.obcResolution.Value);
-            List<int> resolutionList = GetNumberListFromStringList(resolutionStrings);
-            if (resolutionList.Count == 2)
-            {
-                OpenBodyCams.defaultRes = new Vector2Int(resolutionList[0], resolutionList[1]);
-                suitsTerminal.Log.LogInfo($"Resolution set to {resolutionList[0]}x{resolutionList[1]}");
-            }
-            else
-            {
-                OpenBodyCams.defaultRes = new Vector2Int(1000, 700);
-                suitsTerminal.Log.LogInfo($"Unable to set resolution to values provided in config: {SConfig.obcResolution.Value}\nUsing default of 1000x700");
-            }
-                
-                
-        }
     }
 
-
-    [HarmonyPatch(typeof(Terminal), "ParsePlayerSentence")]
-    public class Terminal_ParsePlayerSentence_Patch
-    {
-        static void Postfix(ref TerminalNode __result)
-        {
-            if (CommandStuff.GetNewDisplayText(ref __result))
-                suitsTerminal.X("command found in base funcstring listing...");
-            return;
-        }
-    }
-
-        public class Page
+    public class Page
     {
         public StringBuilder Content { get; set; }
         public int PageNumber { get; set; }