+++ Plugin Core/suitsTerminal.cs
@@ -2,24 +2,28 @@
 using BepInEx.Bootstrap;
 using BepInEx.Logging;
 using HarmonyLib;
+using OpenLib.CoreMethods;
+using suitsTerminal.EventSub;
 using System.Reflection;
 using static suitsTerminal.Misc;
 
 namespace suitsTerminal
 {
-    [BepInPlugin("darmuh.suitsTerminal", "suitsTerminal", "1.3.0")]
+    [BepInPlugin("darmuh.suitsTerminal", "suitsTerminal", PluginInfo.PLUGIN_VERSION)]
+    [BepInDependency("darmuh.OpenLib", "0.1.3")]
 
     public class suitsTerminal : BaseUnityPlugin
     {
         public static suitsTerminal instance;
         internal static bool TerminalStuff = false;
         internal static bool OpenBodyCams = false;
+        //internal static MainListing sT = new();
 
         public static class PluginInfo
         {
             public const string PLUGIN_GUID = "darmuh.suitsTerminal";
             public const string PLUGIN_NAME = "suitsTerminal";
-            public const string PLUGIN_VERSION = "1.3.0";
+            public const string PLUGIN_VERSION = "1.4.0";
         }
 
         public static Terminal Terminal;
@@ -30,15 +34,13 @@ namespace suitsTerminal
         {
             instance = this;
             Log = base.Logger;
-
+            //CommandRegistry.InitListing(ref sT); //using openLib defaultListing
             Log.LogInfo($"{PluginInfo.PLUGIN_NAME} version {PluginInfo.PLUGIN_VERSION} has been started!");
             keywordsCreated = false;
             SConfig.Settings();
-
             Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly());
 
-            //LeaveTerminal.AddTest(); //this command is only for devtesting
-            //Addkeywords used to be here
+            Subscribers.Subscribe();
 
             //start of networking stuff
             /*
@@ -64,7 +66,7 @@ namespace suitsTerminal
             if (!SConfig.extensiveLogging.Value)
                 return;
 
-            suitsTerminal.Log.LogInfo(message);
+            Log.LogInfo(message);
         }
 
     }