+++ Terminal Stuff/OldCommands.cs
@@ -3,9 +3,9 @@ using System;
 using static suitsTerminal.StringStuff;
 using static suitsTerminal.Misc;
 using static suitsTerminal.AllSuits;
-using static suitsTerminal.TerminalHook;
 using System.Collections.Generic;
 using System.Text;
+using OpenLib.ConfigManager;
 
 namespace suitsTerminal
 {
@@ -27,23 +27,15 @@ namespace suitsTerminal
                         suitsTerminal.X($"Duplicate found. Updated SuitName: {SuitName}");
                     }
                     suitNames.Add(SuitName.ToLower());
-                    CommandHandler.AddCommand(SuitName, true, "wear " + SuitName, false, SuitName, CommandHandler.SuitPickCommand, CommandStuff.sT);
+                    CommandHandler.AddCommand(true, "wear " + SuitName, SuitName, CommandHandler.SuitPickCommand, ConfigSetup.defaultListing);
                     suitsTerminal.X($"Keyword for {SuitName} added");
                 }
                 else if (item.syncedSuitID.Value >= 0 && keywordsCreated)
                 {
                     SuitName = UnlockableItems[item.syncedSuitID.Value].unlockableName;
                     SuitName = TerminalFriendlyString(SuitName);
-                    if (TryGetKeyword("wear " + SuitName))
-                    {
-                        CommandHandler.AddCommand(SuitName, true, "wear " + SuitName, false, SuitName, CommandHandler.SuitPickCommand, CommandStuff.sT);
-                        suitsTerminal.X($"Keyword for {SuitName} updated");
-                    }
-                    else
-                    {
-                        CommandHandler.AddCommand(SuitName, true, "wear " + SuitName, false, SuitName, CommandHandler.SuitPickCommand, CommandStuff.sT);
-                        suitsTerminal.X($"Keyword for {SuitName} added, keyword appears to have been removed.");
-                    }
+                    CommandHandler.AddCommand(true, "wear " + SuitName, SuitName, CommandHandler.SuitPickCommand, ConfigSetup.defaultListing);
+                    suitsTerminal.X($"Keyword for {SuitName} updated");
 
                 }
                 else if (item.syncedSuitID.Value < 0)
@@ -146,34 +138,16 @@ namespace suitsTerminal
         // Helper method to create or update the main suit command
         private static void CreateOrUpdateMainSuitCommand(Page page)
         {
-            if (TryGetKeyword("suits"))
-            {
-                //MakeCommand("Quit Terminal", keyword, "", false, true, QuitTerminalCommand, darmuhsTerminalStuff);
-                MakeCommand("suits (main)", "suits", $"{page.Content}", false, true);
+                CommandHandler.AddBasicCommand("suits (main)", "suits", $"{page.Content}", "other", "Display a listing of available suits to wear");
                 suitsTerminal.X($"Updating main suits command");
-            }
-            else
-            {
-                MakeCommand("suits (main)", "suits", $"{page.Content}", false, true);
-                suitsTerminal.X($"main suits command was deleted, creating again");
-            }
         }
 
         // Helper method to create or update suit pages
         private static void CreateOrUpdateSuitPage(Page page)
         {
-            if (TryGetKeyword($"suits {page.PageNumber}"))
-            {
-                suitsTerminal.X($"page {page.PageNumber} has already been creating, updating keyword");
-                MakeCommand($"suits (pg.{page.PageNumber})", $"suits {page.PageNumber}", $"{page.Content}", false, true);
-                //suitsTerminal.X($"Created keyword 'suits {page.PageNumber}'");
-            }
-            else
-            {
-                suitsTerminal.X($"page {page.PageNumber} has been deleted, creating keyword again");
-                MakeCommand($"suits (pg.{page.PageNumber})", $"suits {page.PageNumber}", $"{page.Content}", false, true);
-                //suitsTerminal.X($"Created keyword 'suits {page.PageNumber}'");
-            }
+            suitsTerminal.X($"Creating page {page.PageNumber} keyword");
+            CommandHandler.AddBasicCommand("suits (pg.{page.PageNumber})", $"suits {page.PageNumber}", $"{page.Content}");
+            //suitsTerminal.X($"Created keyword 'suits {page.PageNumber}'");
         }
 
         internal static void MakeRandomSuitCommand()
@@ -181,7 +155,8 @@ namespace suitsTerminal
             if (!SConfig.randomSuitCommand.Value)
                 return;
             
-                CommandHandler.AddCommand("random suit command", true, "randomsuit", false, "sT_random_suit", CommandHandler.RandomSuitCommand, CommandStuff.sT);
+            CommandHandler.AddCommand(true, "randomsuit", "random suit command", CommandHandler.RandomSuit, ConfigSetup.defaultListing, "other", "Equip a random suit");
+
         }
     }
 }